Rename MultiAxisFocusContextPlot to TSAVisualizationCard
